build, cmd, mobile, params, swarm: remove VERSION file; rename Version to VersionMeta;
