--- VERSION
@@ -1,1 +1,0 @@
-1.8.13
