+++ cmd/swarm/config.go
@@ -38,8 +38,6 @@ import (
 	bzzapi "github.com/ethereum/go-ethereum/swarm/api"
 )
 
-const SWARM_VERSION = "0.3.1-unstable"
-
 var (
 	//flag definition for the dumpconfig command
 	DumpConfigCommand = cli.Command{
