--- swarm/VERSION
@@ -1,1 +1,0 @@
-0.3.1
